+++ test/test_listener.dart
@@ -1,0 +1,9 @@
+class TestListener {
+  TestListener();
+
+  bool wasCalled = false;
+
+  void notifyListeners() {
+    wasCalled = true;
+  }
+}
